accrual_create_df

Creates a data frame that contains the absolute and cumululative number of patients recruited at each date from a vector with enrollment dates.

accrual_create_df(
  enrollment_dates,
  format_enrollment_dates = "%d%b%Y",
  start_date = NA,
  format_start_date = "%d%b%Y",
  current_date = NA,
  format_current_date = "%d%b%Y",
  by = NA,
  overall = TRUE,
  name_overall = "Overall"
)

Arguments

enrollment_dates
dates on which patients are enrolled as date or character vector

format_enrollment_dates
single character identifying the format of the dates, ignored if enrollment_dates are dates

start_date
date when recruitment started, single character or date, if not given the first enrollment date is used

format_start_date
format of the start date, ignored if start_date is a date

current_date
date of the data export or database freeze, single character or date, if not given the latest enrollment date is used

format_current_date
format of the current date, ignored if current_date is a date

by
vector with centers, has to have the same length as enrollment dates, generates a list with accrual data frames for each site

overall
indicates that accrual_df contains a summary with all sites (only if by is not NA)

name_overall
name of the summary with all sites (if by is not NA and overall==TRUE)

Value

Returns a data frame (or a list of data frames if by is not NA) with three columns "Date", "Freq" and "Cumulative" with each date with an accural and the absolute and cumulative number of patients accrued.

Examples

# \donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_create_df(enrollment_dates)

#>          Date Freq Cumulative
#> 1  2018-01-02    1          1
#> 2  2018-01-03    2          3
#> 3  2018-01-04    1          4
#> 4  2018-01-05    2          6
#> 5  2018-01-07    2          8
#> 6  2018-01-08    1          9
#> 7  2018-01-09    2         11
#> 8  2018-01-11    3         14
#> 9  2018-01-12    1         15
#> 10 2018-01-13    3         18
#> 11 2018-01-14    2         20
#> 12 2018-01-15    3         23
#> 13 2018-01-16    2         25
#> 14 2018-01-17    4         29
#> 15 2018-01-18    2         31
#> 16 2018-01-19    4         35
#> 17 2018-01-21    3         38
#> 18 2018-01-22    1         39
#> 19 2018-01-23    3         42
#> 20 2018-01-24    1         43
#> 21 2018-01-25    3         46
#> 22 2018-01-28    1         47
#> 23 2018-01-29    1         48
#> 24 2018-01-30    2         50

# different start and current date
accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
current_date=as.Date("2018-03-01"))

#>          Date Freq Cumulative
#> 1  2017-12-01    0          0
#> 2  2018-01-02    1          1
#> 3  2018-01-03    2          3
#> 4  2018-01-04    1          4
#> 5  2018-01-05    2          6
#> 6  2018-01-07    2          8
#> 7  2018-01-08    1          9
#> 8  2018-01-09    2         11
#> 9  2018-01-11    3         14
#> 10 2018-01-12    1         15
#> 11 2018-01-13    3         18
#> 12 2018-01-14    2         20
#> 13 2018-01-15    3         23
#> 14 2018-01-16    2         25
#> 15 2018-01-17    4         29
#> 16 2018-01-18    2         31
#> 17 2018-01-19    4         35
#> 18 2018-01-21    3         38
#> 19 2018-01-22    1         39
#> 20 2018-01-23    3         42
#> 21 2018-01-24    1         43
#> 22 2018-01-25    3         46
#> 23 2018-01-28    1         47
#> 24 2018-01-29    1         48
#> 25 2018-01-30    2         50
#> 26 2018-03-01    0         50

#by site
set.seed(2020)
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_create_df(enrollment_dates,by=centers)

#> $`Site 3`
#>          Date Freq Cumulative
#> 1  2018-01-02    1          1
#> 2  2018-01-11    1          2
#> 3  2018-01-14    1          3
#> 4  2018-01-15    1          4
#> 5  2018-01-16    1          5
#> 6  2018-01-17    1          6
#> 7  2018-01-18    1          7
#> 8  2018-01-19    1          8
#> 9  2018-01-21    1          9
#> 10 2018-01-23    1         10
#> 11 2018-01-25    1         11
#> 12 2018-01-29    1         12
#> 
#> $`Site 2`
#>          Date Freq Cumulative
#> 1  2018-01-03    1          1
#> 2  2018-01-05    2          3
#> 3  2018-01-08    1          4
#> 4  2018-01-09    2          6
#> 5  2018-01-11    1          7
#> 6  2018-01-12    1          8
#> 7  2018-01-13    2         10
#> 8  2018-01-14    1         11
#> 9  2018-01-15    2         13
#> 10 2018-01-17    3         16
#> 11 2018-01-18    1         17
#> 12 2018-01-21    1         18
#> 13 2018-01-23    1         19
#> 14 2018-01-30    2         21
#> 
#> $`Site 1`
#>          Date Freq Cumulative
#> 1  2018-01-03    1          1
#> 2  2018-01-04    1          2
#> 3  2018-01-07    2          4
#> 4  2018-01-11    1          5
#> 5  2018-01-13    1          6
#> 6  2018-01-16    1          7
#> 7  2018-01-19    3         10
#> 8  2018-01-21    1         11
#> 9  2018-01-22    1         12
#> 10 2018-01-23    1         13
#> 11 2018-01-24    1         14
#> 12 2018-01-25    2         16
#> 13 2018-01-28    1         17
#> 
#> $Overall
#>          Date Freq Cumulative
#> 1  2018-01-02    1          1
#> 2  2018-01-03    2          3
#> 3  2018-01-04    1          4
#> 4  2018-01-05    2          6
#> 5  2018-01-07    2          8
#> 6  2018-01-08    1          9
#> 7  2018-01-09    2         11
#> 8  2018-01-11    3         14
#> 9  2018-01-12    1         15
#> 10 2018-01-13    3         18
#> 11 2018-01-14    2         20
#> 12 2018-01-15    3         23
#> 13 2018-01-16    2         25
#> 14 2018-01-17    4         29
#> 15 2018-01-18    2         31
#> 16 2018-01-19    4         35
#> 17 2018-01-21    3         38
#> 18 2018-01-22    1         39
#> 19 2018-01-23    3         42
#> 20 2018-01-24    1         43
#> 21 2018-01-25    3         46
#> 22 2018-01-28    1         47
#> 23 2018-01-29    1         48
#> 24 2018-01-30    2         50
#> 
#> attr(,"class")
#> [1] "accrual_df" "list"      

# }