accrual_create_df

Creates a data frame that contains the absolute and cumululative number of patients recruited at each date from a vector with enrollment dates.

accrual_create_df(
  enrollment_dates,
  start_date = "site",
  current_date = "common",
  force_start0 = TRUE,
  by = NA,
  overall = TRUE,
  name_overall = "Overall",
  pos_overall = c("last", "first")
)

Arguments

enrollment_dates
dates on which patients are enrolled, date vector

start_date
date when recruitment started. Single date (used for all sites in by), date vector (with the same length the number of distinct sites in by), "common" (first date overall) or "site" (first date for each site, default).

current_date
date of the data export or database freeze. Single date, date vector (with the same length the number of distinct sites in by), "common" (last date overall, default) or "site" (first date for each site).

force_start0
logical, adds an extra 0 line to the accrual data frame in cases where a start date is given and corresponds to the earliest enrollment date.

by
vector with sites, has to have the same length as enrollment dates, generates a list with accrual data frames for each site

overall
logical indicates that accrual_df contains a summary with all sites (only if by is not NA)

name_overall
name of the summary with all sites (if by is not NA and overall==TRUE)

pos_overall
overall as last or first element of the list (if by is not NA and overall==TRUE)

Value

Returns a data frame (or a list of data frames if by is not NA) with three columns "Date", "Freq" and "Cumulative" with each date with an accrual and the absolute and cumulative number of patients accrued.

Examples

# \donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_create_df(enrollment_dates)
#> 50 participants recruited between 2018-01-02 and 2018-01-30 
#>         Date Freq Cumulative
#> 1 2018-01-02    0          0
#> 2 2018-01-02    1          1
#> 3 2018-01-03    2          3
#> 4 2018-01-04    1          4
#> 5 2018-01-05    2          6
#> 6 2018-01-07    2          8
# different start and current date
accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
current_date=as.Date("2018-03-01"))
#> 50 participants recruited between 2017-12-01 and 2018-03-01 
#>         Date Freq Cumulative
#> 1 2017-12-01    0          0
#> 2 2018-01-02    1          1
#> 3 2018-01-03    2          3
#> 4 2018-01-04    1          4
#> 5 2018-01-05    2          6
#> 6 2018-01-07    2          8
#by site
set.seed(2020)
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_create_df(enrollment_dates,by=centers)
#> 17 participants recruited between 2018-01-03 and 2018-01-30 
#>         Date Freq Cumulative
#> 1 2018-01-03    0          0
#> 2 2018-01-03    1          1
#> 3 2018-01-04    1          2
#> 4 2018-01-07    2          4
#> 5 2018-01-11    1          5
#> 6 2018-01-13    1          6
#> 21 participants recruited between 2018-01-03 and 2018-01-30 
#>         Date Freq Cumulative
#> 1 2018-01-03    0          0
#> 2 2018-01-03    1          1
#> 3 2018-01-05    2          3
#> 4 2018-01-08    1          4
#> 5 2018-01-09    2          6
#> 6 2018-01-11    1          7
#> 12 participants recruited between 2018-01-02 and 2018-01-30 
#>         Date Freq Cumulative
#> 1 2018-01-02    0          0
#> 2 2018-01-02    1          1
#> 3 2018-01-11    1          2
#> 4 2018-01-14    1          3
#> 5 2018-01-15    1          4
#> 6 2018-01-16    1          5
#> 50 participants recruited between 2018-01-02 and 2018-01-30 
#>         Date Freq Cumulative
#> 1 2018-01-02    0          0
#> 2 2018-01-02    1          1
#> 3 2018-01-03    2          3
#> 4 2018-01-04    1          4
#> 5 2018-01-05    2          6
#> 6 2018-01-07    2          8
# }