accrual_table

Table of recruitment overview by site, rate of recruitment

accrual_table(
  accrual_df,
  overall = TRUE,
  name_overall = "Overall",
  pos_overall = c("last", "first"),
  unit = c("month", "year", "week", "day"),
  format_table_date = "%d%b%Y",
  format_time = "%1.0f",
  format_rrate = "%1.2f",
  header = TRUE
)

Arguments

accrual_df

object of class 'accrual_df' or 'accrual_list' produced by accrual_create_df.

overall

logical, indicates that accrual_df contains a summary with all sites (only if by is not NA).

name_overall

name of the summary with all sites (if by is not NA and overall==TRUE).

pos_overall

overall in last or first row (if by is not NA and overall==TRUE).

unit

time unit for time recruiting and the rate, one of "month", "year", "week" or "day".

format_table_date

format of start date in table.

format_time

format of time recruiting in table.

format_rrate

format of recruitment rate in table.

header

include header, logical or character vector of length 4 or 5 (if accrual_df is a list).

Value

Returns data frame with a header, a row per site and overall and the following columns:

name

name of the site (if accrual_df is a list)

start_date

accrual start date

time

time accruing

n

number of patients accrued

rate

accrual rate per time unit

Examples

set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_df<-accrual_create_df(enrollment_dates,by=centers)
accrual_table(accrual_df)
#>      name           start_date            time                    n
#> 1  Center First participant in Months accruing Participants accrued
#> 2  Site 1            02Jan2018               1                   19
#> 3  Site 2            03Jan2018               1                   19
#> 4  Site 3            04Jan2018               1                   12
#> 5 Overall            02Jan2018               1                   50
#>                       rate
#> 1 Accrual rate (per month)
#> 2                    20.36
#> 3                    21.11
#> 4                    13.85
#> 5                    53.57

#format
accrual_table(accrual_df,format_time="%1.1f",format_rrate="%1.1f")
#>      name           start_date            time                    n
#> 1  Center First participant in Months accruing Participants accrued
#> 2  Site 1            02Jan2018             0.9                   19
#> 3  Site 2            03Jan2018             0.9                   19
#> 4  Site 3            04Jan2018             0.9                   12
#> 5 Overall            02Jan2018             0.9                   50
#>                       rate
#> 1 Accrual rate (per month)
#> 2                     20.4
#> 3                     21.1
#> 4                     13.8
#> 5                     53.6

#unit
accrual_table(accrual_df,unit="day")
#>      name           start_date          time                    n
#> 1  Center First participant in Days accruing Participants accrued
#> 2  Site 1            02Jan2018            28                   19
#> 3  Site 2            03Jan2018            27                   19
#> 4  Site 3            04Jan2018            26                   12
#> 5 Overall            02Jan2018            28                   50
#>                     rate
#> 1 Accrual rate (per day)
#> 2                   0.68
#> 3                   0.70
#> 4                   0.46
#> 5                   1.79

#common start and current dates
accrual_df<-accrual_create_df(enrollment_dates,by=centers,start_date="common",current_date="common")
accrual_table(accrual_df)
#>      name           start_date            time                    n
#> 1  Center First participant in Months accruing Participants accrued
#> 2  Site 1            02Jan2018               1                   19
#> 3  Site 2            02Jan2018               1                   19
#> 4  Site 3            02Jan2018               1                   12
#> 5 Overall            02Jan2018               1                   50
#>                       rate
#> 1 Accrual rate (per month)
#> 2                    20.36
#> 3                    20.36
#> 4                    12.86
#> 5                    53.57
accrual_df<-accrual_create_df(enrollment_dates,by=centers,start_date=as.Date("2017-12-31"),
    current_date=as.Date("2018-03-01"))
accrual_table(accrual_df)
#>      name           start_date            time                    n
#> 1  Center First participant in Months accruing Participants accrued
#> 2  Site 1            31Dec2017               2                   19
#> 3  Site 2            31Dec2017               2                   19
#> 4  Site 3            31Dec2017               2                   12
#> 5 Overall            31Dec2017               2                   50
#>                       rate
#> 1 Accrual rate (per month)
#> 2                     9.50
#> 3                     9.50
#> 4                     6.00
#> 5                    25.00