accrual_table

Table of recruitment overview by site, including name of site, start date, time recruiting, number of patients recruited and rate of recruitment

accrual_table(
  accrual_df,
  overall = TRUE,
  name_overall = "Overall",
  pos_overall = c("last", "first"),
  start_date = NA,
  format_start_date = "%d%b%Y",
  current_date = NA,
  format_current_date = "%d%b%Y",
  unit = c("month", "year", "week", "day"),
  format_table_date = "%d%b%Y",
  format_time = "%1.0f",
  format_rrate = "%1.2f",
  header = TRUE
)

Arguments

accrual_df
accrual data frame produced by accrual_create_df potentially with by option (i.e. as a list)

overall
logical, indicates that accrual_df contains a summary with all sites (only if by is not NA)

name_overall
name of the summary with all sites (if by is not NA and overall==TRUE)

pos_overall
overall in last or first row (if by is not NA and overall==TRUE)

start_date
start_date: date when recruitment started, single character or date,

format_start_date
format of the start date, ignored if start_date is a date

current_date
date of the data export or database freeze, single character or date

format_current_date
format of the current date, ignored if current_date is a date

unit
time unit for time recruiting and the rate, any of "month","year","week","day"

format_table_date
format of start date in table

format_time
format of time recruiting in table

format_rrate
format of recruitment rate in table

header
header: include header, TRUE, NULL or character vector of length 5

Value

Data frame with name of the site, accrual start date, time accruing, number of patients accrued and accrual rate.

Examples

set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_df<-accrual_create_df(enrollment_dates,by=centers)
accrual_table(accrual_df)
#>         name           start_date            time                    n
#> 1     Center First participant in Months accruing Participants accrued
#> tabi  Site 1            02Jan2018               1                   19
#> 2     Site 2            03Jan2018               1                   19
#> 3     Site 3            04Jan2018               1                   12
#> 4    Overall            02Jan2018               1                   50
#>                          rate
#> 1    Accrual rate (per month)
#> tabi                    20.36
#> 2                       21.11
#> 3                       13.85
#> 4                       53.57

#format
accrual_table(accrual_df,format_time="%1.1f",format_rrate="%1.1f")
#>         name           start_date            time                    n
#> 1     Center First participant in Months accruing Participants accrued
#> tabi  Site 1            02Jan2018             0.9                   19
#> 2     Site 2            03Jan2018             0.9                   19
#> 3     Site 3            04Jan2018             0.9                   12
#> 4    Overall            02Jan2018             0.9                   50
#>                          rate
#> 1    Accrual rate (per month)
#> tabi                     20.4
#> 2                        21.1
#> 3                        13.8
#> 4                        53.6

#unit
accrual_table(accrual_df,unit="day")
#>         name           start_date          time                    n
#> 1     Center First participant in Days accruing Participants accrued
#> tabi  Site 1            02Jan2018            28                   19
#> 2     Site 2            03Jan2018            27                   19
#> 3     Site 3            04Jan2018            26                   12
#> 4    Overall            02Jan2018            28                   50
#>                        rate
#> 1    Accrual rate (per day)
#> tabi                   0.68
#> 2                      0.70
#> 3                      0.46
#> 4                      1.79

#common start and current dates
accrual_table(accrual_df,unit="day",start_date="common",current_date="common")
#>         name           start_date          time                    n
#> 1     Center First participant in Days accruing Participants accrued
#> tabi  Site 1            02Jan2018            28                   19
#> 2     Site 2            02Jan2018            28                   19
#> 3     Site 3            02Jan2018            28                   12
#> 4    Overall            02Jan2018            28                   50
#>                        rate
#> 1    Accrual rate (per day)
#> tabi                   0.68
#> 2                      0.68
#> 3                      0.43
#> 4                      1.79
accrual_table(accrual_df,unit="day",start_date=as.Date("2017-12-31"),
    current_date=as.Date("2018-03-01"))
#>         name           start_date          time                    n
#> 1     Center First participant in Days accruing Participants accrued
#> tabi  Site 1            31Dec2017            60                   19
#> 2     Site 2            31Dec2017            60                   19
#> 3     Site 3            31Dec2017            60                   12
#> 4    Overall            31Dec2017            60                   50
#>                        rate
#> 1    Accrual rate (per day)
#> tabi                   0.32
#> 2                      0.32
#> 3                      0.20
#> 4                      0.83