accrual_time_unit

Generates summary of recruitment per time unit

accrual_time_unit(accrual_df, unit = c("month", "year", "week", "day"))

Arguments

accrual_df

accrual data frame produced by accrual_create_df with by=NA.

unit

time unit for which the bars should be plotted, one of "month", "year", "week" or "day".

Value

Returns a data frame with the number of patients accrued for each time unit.

Examples

# \donttest{
data(accrualdemo)
accrual_df<-accrual_create_df(accrualdemo$date)
accrual_time_unit(accrual_df,"week")
#>    week       date Freq
#> 1     1 2020-07-06    1
#> 2     2 2020-07-13    5
#> 3     3 2020-07-20   10
#> 4     4 2020-07-27    9
#> 5     5 2020-08-03   12
#> 6     6 2020-08-10   13
#> 7     7 2020-08-17   14
#> 8     8 2020-08-24   24
#> 9     9 2020-08-31   22
#> 10   10 2020-09-07   24
#> 11   11 2020-09-14   36
#> 12   12 2020-09-21   33
#> 13   13 2020-09-28   33
#> 14   14 2020-10-05    3
accrual_time_unit(accrual_df,"day")
#>    day month year Freq       date
#> 1    9     7 2020    1 2020-07-09
#> 80  10     7 2020    0 2020-07-10
#> 81  11     7 2020    0 2020-07-11
#> 82  12     7 2020    0 2020-07-12
#> 79  13     7 2020    0 2020-07-13
#> 2   14     7 2020    3 2020-07-14
#> 85  15     7 2020    0 2020-07-15
#> 3   16     7 2020    1 2020-07-16
#> 83  17     7 2020    0 2020-07-17
#> 88  18     7 2020    0 2020-07-18
#> 4   19     7 2020    1 2020-07-19
#> 5   20     7 2020    1 2020-07-20
#> 6   21     7 2020    2 2020-07-21
#> 7   22     7 2020    1 2020-07-22
#> 8   23     7 2020    4 2020-07-23
#> 90  24     7 2020    0 2020-07-24
#> 9   25     7 2020    1 2020-07-25
#> 10  26     7 2020    1 2020-07-26
#> 11  27     7 2020    2 2020-07-27
#> 93  28     7 2020    0 2020-07-28
#> 86  29     7 2020    0 2020-07-29
#> 12  30     7 2020    1 2020-07-30
#> 13  31     7 2020    3 2020-07-31
#> 89   1     8 2020    0 2020-08-01
#> 14   2     8 2020    3 2020-08-02
#> 15   3     8 2020    2 2020-08-03
#> 16   4     8 2020    5 2020-08-04
#> 17   5     8 2020    2 2020-08-05
#> 18   6     8 2020    1 2020-08-06
#> 87   7     8 2020    0 2020-08-07
#> 19   8     8 2020    2 2020-08-08
#> 84   9     8 2020    0 2020-08-09
#> 20  10     8 2020    2 2020-08-10
#> 21  11     8 2020    1 2020-08-11
#> 22  12     8 2020    1 2020-08-12
#> 23  13     8 2020    2 2020-08-13
#> 24  14     8 2020    3 2020-08-14
#> 25  15     8 2020    1 2020-08-15
#> 26  16     8 2020    3 2020-08-16
#> 27  17     8 2020    4 2020-08-17
#> 92  18     8 2020    0 2020-08-18
#> 91  19     8 2020    0 2020-08-19
#> 28  20     8 2020    3 2020-08-20
#> 29  21     8 2020    1 2020-08-21
#> 30  22     8 2020    3 2020-08-22
#> 31  23     8 2020    3 2020-08-23
#> 32  24     8 2020    1 2020-08-24
#> 33  25     8 2020    4 2020-08-25
#> 34  26     8 2020    4 2020-08-26
#> 35  27     8 2020    2 2020-08-27
#> 36  28     8 2020    4 2020-08-28
#> 37  29     8 2020    6 2020-08-29
#> 38  30     8 2020    3 2020-08-30
#> 39  31     8 2020    3 2020-08-31
#> 40   1     9 2020    5 2020-09-01
#> 41   2     9 2020    2 2020-09-02
#> 42   3     9 2020    4 2020-09-03
#> 43   4     9 2020    2 2020-09-04
#> 44   5     9 2020    3 2020-09-05
#> 45   6     9 2020    3 2020-09-06
#> 46   7     9 2020    5 2020-09-07
#> 47   8     9 2020    3 2020-09-08
#> 48   9     9 2020    2 2020-09-09
#> 49  10     9 2020    2 2020-09-10
#> 50  11     9 2020    2 2020-09-11
#> 51  12     9 2020    6 2020-09-12
#> 52  13     9 2020    4 2020-09-13
#> 53  14     9 2020    6 2020-09-14
#> 54  15     9 2020    2 2020-09-15
#> 55  16     9 2020    8 2020-09-16
#> 56  17     9 2020    6 2020-09-17
#> 57  18     9 2020    4 2020-09-18
#> 58  19     9 2020    6 2020-09-19
#> 59  20     9 2020    4 2020-09-20
#> 60  21     9 2020    2 2020-09-21
#> 61  22     9 2020    6 2020-09-22
#> 62  23     9 2020    4 2020-09-23
#> 63  24     9 2020    4 2020-09-24
#> 64  25     9 2020    6 2020-09-25
#> 65  26     9 2020    5 2020-09-26
#> 66  27     9 2020    6 2020-09-27
#> 67  28     9 2020    7 2020-09-28
#> 68  29     9 2020    2 2020-09-29
#> 69  30     9 2020    4 2020-09-30
#> 70   1    10 2020    4 2020-10-01
#> 71   2    10 2020    9 2020-10-02
#> 72   3    10 2020    3 2020-10-03
#> 73   4    10 2020    4 2020-10-04
#> 74   5    10 2020    3 2020-10-05
#> 75   6    10 2020    2 2020-10-06
#> 76   7    10 2020    6 2020-10-07
#> 77   8    10 2020    1 2020-10-08
#> 78   9    10 2020    2 2020-10-09
# }