accrual_time_unit

Generates summary of recruitment per time unit

accrual_time_unit(accrual_df, unit = c("month", "year", "week", "day"))

Arguments

accrual_df
accrual data frame produced by accrual_create_df with by=NA

unit
time unit for which the bars should be plotted, any of "month","year","week","day"

Value

Data frame with the number of patients accrued for each time unit (given by day and/or week and/or month and year).

Examples

# \donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrualPlot:::accrual_time_unit(accrual_df,"week")
#>   week month year Freq       date
#> 1    1     1 2018    8 2018-01-01
#> 2    2     1 2018   12 2018-01-08
#> 3    3     1 2018   18 2018-01-15
#> 4    4     1 2018    9 2018-01-22
#> 5    5     1 2018    3 2018-01-29
accrualPlot:::accrual_time_unit(accrual_df,"day")
#>    day month year Freq       date
#> 1    2     1 2018    1 2018-01-02
#> 2    3     1 2018    2 2018-01-03
#> 3    4     1 2018    1 2018-01-04
#> 4    5     1 2018    2 2018-01-05
#> 25   6     1 2018    0 2018-01-06
#> 5    7     1 2018    2 2018-01-07
#> 6    8     1 2018    1 2018-01-08
#> 7    9     1 2018    2 2018-01-09
#> 28  10     1 2018    0 2018-01-10
#> 8   11     1 2018    3 2018-01-11
#> 9   12     1 2018    1 2018-01-12
#> 10  13     1 2018    3 2018-01-13
#> 11  14     1 2018    2 2018-01-14
#> 12  15     1 2018    3 2018-01-15
#> 13  16     1 2018    2 2018-01-16
#> 14  17     1 2018    4 2018-01-17
#> 15  18     1 2018    2 2018-01-18
#> 16  19     1 2018    4 2018-01-19
#> 27  20     1 2018    0 2018-01-20
#> 17  21     1 2018    3 2018-01-21
#> 18  22     1 2018    1 2018-01-22
#> 19  23     1 2018    3 2018-01-23
#> 20  24     1 2018    1 2018-01-24
#> 21  25     1 2018    3 2018-01-25
#> 29  26     1 2018    0 2018-01-26
#> 26  27     1 2018    0 2018-01-27
#> 22  28     1 2018    1 2018-01-28
#> 23  29     1 2018    1 2018-01-29
#> 24  30     1 2018    2 2018-01-30
# }